Move test support into conftest and helpers
