+++ tests/conftest.py
@@ -2,6 +2,8 @@ import tempfile
 
 import pytest
 from flask import Flask
+import mock
+from boto.ec2.cloudwatch import CloudWatchConnection
 
 from dmutils.logging import init_app
 
@@ -17,3 +19,19 @@ def app_with_logging(app):
         app.config['DM_LOG_PATH'] = f.name
         init_app(app)
         yield app
+
+
+@pytest.yield_fixture
+def cloudwatch():
+    with mock.patch('dmutils.metrics.connect_to_region') as connect_to_region:
+        conn = mock.Mock(spec=CloudWatchConnection)
+        connect_to_region.return_value = conn
+        yield conn
+
+
+@pytest.fixture
+def os_environ(request):
+    env_patch = mock.patch('os.environ', {})
+    request.addfinalizer(env_patch.stop)
+
+    return env_patch.start()
