+++ tests/helpers.py
@@ -1,0 +1,6 @@
+from datetime import datetime
+
+
+class IsDatetime(object):
+    def __eq__(self, other):
+        return isinstance(other, datetime)
